+++ GameEngine2.vcxproj
@@ -155,19 +155,18 @@
     <ClCompile Include="dependencies\include\imgui\imgui_widgets.cpp" />
     <ClCompile Include="dependencies\include\jsoncpp.cpp" />
     <ClCompile Include="engine\Application\Application.cpp" />
-    <ClCompile Include="engine\Controller\Emotion.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\Emotion.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\AffordanceController.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Pickup.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Poke.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Punch.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Slap.cpp" />
-    <ClCompile Include="engine\Controller\EmotionManager.cpp" />
     <ClCompile Include="engine\Controller\FCM.cpp" />
-    <ClCompile Include="engine\Controller\OCCModel.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\OCCModel.cpp" />
     <ClCompile Include="engine\Controller\AI\Pathfinding\A_Star.cpp" />
     <ClCompile Include="engine\Controller\AI\Pathfinding\PathNode.cpp" />
     <ClCompile Include="engine\Controller\AI\Pathfinding\PathNodeManager.cpp" />
-    <ClCompile Include="engine\Controller\Personality.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\Personality.cpp" />
     <ClCompile Include="engine\Controller\Utils\DebugLogger.cpp" />
     <ClCompile Include="engine\Controller\Physics\CollisionSolver.cpp" />
     <ClCompile Include="engine\Controller\Event\EventDispatcher.cpp" />
@@ -243,8 +242,8 @@
     <ClInclude Include="engine\Controller\AI\Pathfinding\PathNode.h" />
     <ClInclude Include="engine\Controller\AI\Pathfinding\PathNodeManager.h" />
     <ClInclude Include="engine\Controller\AudioDataTypes.h" />
-    <ClInclude Include="engine\Controller\Emotion.h" />
-    <ClInclude Include="engine\Controller\Personality.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\Emotion.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\Personality.h" />
     <ClInclude Include="engine\Controller\Utils\DebugLogger.h" />
     <ClInclude Include="engine\Controller\Utils\StreamCapture.h" />
     <ClInclude Include="engine\Controller\Event\Event.h" />
@@ -255,7 +254,7 @@
     <ClInclude Include="engine\Controller\Lua\LuaGameBridge.h" />
     <ClInclude Include="engine\Controller\Lua\LuaManager.h" />
     <ClInclude Include="Engine\Controller\md2Reader.h" />
-    <ClInclude Include="engine\Controller\OCCModel.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\OCCModel.h" />
     <ClInclude Include="Engine\Controller\Physics\Collider.h" />
     <ClInclude Include="engine\Controller\Physics\Collisionsolver.h" />
     <ClInclude Include="engine\Controller\Physics\PhysicsIntegrator.h" />
