+++ GameEngine2.vcxproj.filters
@@ -55,14 +55,12 @@
     <ClCompile Include="Engine\Controller\Serialization\SceneLoader.cpp" />
     <ClCompile Include="engine\Controller\Utils\DebugLogger.cpp" />
     <ClCompile Include="engine\Controller\FCM.cpp" />
-    <ClCompile Include="engine\Controller\OCCModel.cpp" />
-    <ClCompile Include="engine\Controller\Emotion.cpp" />
-    <ClCompile Include="engine\Controller\Personality.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\OCCModel.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\Emotion.cpp" />
+    <ClCompile Include="engine\Controller\AI\Emotion\Personality.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\AffordanceController.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Pickup.cpp" />
-    <ClCompile Include="engine\Controller\EmotionManager.cpp" />
     <ClCompile Include="engine\Controller\FCM.cpp" />
-    <ClCompile Include="engine\Controller\OCCModel.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Poke.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Punch.cpp" />
     <ClCompile Include="engine\Controller\AI\Affordance\Affordance_Slap.cpp" />
@@ -145,15 +143,13 @@
     <ClInclude Include="engine\Controller\Utils\StreamCapture.h" />
     <ClInclude Include="engine\Controller\Utils\DebugLogger.h" />
     <ClInclude Include="engine\Controller\FCM.h" />
-    <ClInclude Include="engine\Controller\OCCModel.h" />
-    <ClInclude Include="engine\Controller\Emotion.h" />
-    <ClInclude Include="engine\Controller\Personality.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\OCCModel.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\Emotion.h" />
+    <ClInclude Include="engine\Controller\AI\Emotion\Personality.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\Affordance.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\AffordanceController.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\Affordance_Pickup.h" />
-    <ClInclude Include="engine\Controller\EmotionManger.h" />
     <ClInclude Include="engine\Controller\FCM.h" />
-    <ClInclude Include="engine\Controller\OCCModel.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\Global_Affordances.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\Affordance_Poke.h" />
     <ClInclude Include="engine\Controller\AI\Affordance\Affordance_Punch.h" />
